+++ src/providers/vero.js
@@ -18,7 +18,7 @@ module.exports = Provider.extend({
   initialize : function (options, ready) {
     window._veroq = window._veroq || [];
     window._veroq.push(['init', { api_key: options.apiKey }]);
-    load('//www.getvero.com/assets/m.js');
+    load('//d3qxef4rp70elm.cloudfront.net/m.js');
 
     // Vero creates a queue, so it's ready immediately.
     ready();