+++ src/providers/visual-website-optimizer.js
@@ -1,61 +1,119 @@
 // http://v2.visualwebsiteoptimizer.com/tools/get_tracking_code.php
+// http://visualwebsiteoptimizer.com/knowledge/integration-of-vwo-with-kissmetrics/
 
-var each      = require('each')
-  , nextTick  = require('next-tick')
-  , Provider  = require('../provider');
+var each = require('each')
+  , inherit = require('inherit')
+  , nextTick = require('next-tick')
+  , Provider = require('../provider');
 
 
-module.exports = Provider.extend({
+/**
+ * Expose `VWO`.
+ */
 
-  name : 'Visual Website Optimizer',
+module.exports = VWO;
 
-  defaults : {
-    // Whether to replay variations into other enabled integrations as traits.
-    variations : true
-  },
 
-  initialize : function (options, ready, analytics) {
-    // Create the `optimizely` object in case it doesn't exist already.
-    // https://www.optimizely.com/docs/api#function-calls
-    window.optimizely = window.optimizely || [];
+/**
+ * `VWO` inherits from the generic `Provider`.
+ */
 
-    // If the `variations` option is true, replay our variations on the next
-    // tick to wait for the entire library to be ready for replays.
-    if (options.variations) {
-      var self = this;
-      nextTick(function () { self.replay(); });
-    }
+function VWO () {
+  Provider.apply(this, arguments);
+}
 
-    // Optimizely should be on the page already, so it's always ready.
-    ready();
-  },
+inherit(VWO, Provider);
 
-  track : function (event, properties) {
-    // Optimizely takes revenue as cents, not dollars.
-    if (properties && properties.revenue) properties.revenue = properties.revenue * 100;
 
-    window.optimizely.push(['trackEvent', event, properties]);
-  },
+/**
+ * Name.
+ */
 
-  replay : function () {
-    // Make sure we have access to Optimizely's `data` dictionary.
-    var data = window.optimizely.data;
-    if (!data) return;
+VWO.prototype.name = 'Visual Website Optimizer';
 
-    // Grab a few pieces of data we'll need for replaying.
-    var experiments       = data.experiments
-      , variationNamesMap = data.state.variationNamesMap;
 
-    // Create our traits object to add variations to.
-    var traits = {};
+/**
+ * Default options.
+ */
 
-    // Loop through all the experiement the user has been assigned a variation
-    // for and add them to our traits.
-    each(variationNamesMap, function (experimentId, variation) {
-      traits['Experiment: ' + experiments[experimentId].name] = variation;
+VWO.prototype.defaults = {
+  // Whether to replay variations into other integrations as traits.
+  replay : true
+};
+
+
+/**
+ * Initialize.
+ */
+
+VWO.prototype.initialize = function (options, ready) {
+  if (options.replay) this.replay();
+  ready();
+};
+
+
+/**
+ * Replay the experiments the user has seen as traits to all other integrations.
+ * Wait for the next tick to replay so that the `analytics` object and all of
+ * the integrations are fully initialized.
+ */
+
+VWO.prototype.replay = function () {
+  var analytics = this.analytics;
+  nextTick(function () {
+    experiments(function (err, traits) {
+      if (traits) analytics.identify(traits);
     });
+  });
+};
 
-    this.analytics.identify(traits);
-  }
 
-});
+/**
+ * Get dictionary of experiment keys and variations.
+ * http://visualwebsiteoptimizer.com/knowledge/integration-of-vwo-with-kissmetrics/
+ *
+ * @param  {Function} callback  Called with `err, experiments`.
+ * @return {Object}             Dictionary of experiments and variations.
+ */
+
+function experiments (callback) {
+  enqueue(function () {
+    var data = {};
+    var ids = window._vwo_exp_ids;
+    if (!ids) return callback();
+    each(ids, function (id) {
+      var name = variation(id);
+      if (name) data['Experiment: ' + id] = name;
+    });
+    callback(null, data);
+  });
+}
+
+
+/**
+ * Add a function to the VWO queue, creating one if it doesn't exist.
+ *
+ * @param {Function} fn  Function to enqueue.
+ */
+
+function enqueue (fn) {
+  window._vis_opt_queue || (window._vis_opt_queue = []);
+  window._vis_opt_queue.push(fn);
+}
+
+
+/**
+ * Get the chosen variation's name from an experiment `id`.
+ * http://visualwebsiteoptimizer.com/knowledge/integration-of-vwo-with-kissmetrics/
+ *
+ * @param  {String} id  ID of the experiment to read.
+ * @return {String}     Variation name.
+ */
+
+function variation (id) {
+  var experiments = window._vwo_exp;
+  if (!experiments) return null;
+  var experiment = experiments[id];
+  var variationId = experiment.combination_chosen;
+  return variationId ? experiment.comb_n[variationId] : null;
+}