+++ test/providers/uservoice.js
@@ -23,8 +23,6 @@ describe('UserVoice', function () {
         primary_color     : '#cc6d00',
         link_color        : '#007dbf',
         default_mode      : 'support',
-        support_tab_name  : null,
-        feedback_tab_name : null,
         tab_label         : 'Feedback & Support',
         tab_color         : '#cc6d00',
         tab_position      : 'middle-right',
@@ -51,17 +49,12 @@ describe('UserVoice', function () {
 
 
   describe('identify', function () {
-
     var extend = require('segmentio-extend');
 
     it('should call setCustomFields', function () {
       var stub = sinon.stub(window.UserVoice, 'push');
-
-      analytics.identify(test.userId, test.traits);
-      expect(stub.called).to.be(true);
-      expect(stub.calledWith(['setCustomFields', extend({}, test.traits, { id : test.userId })])).to.be(true);
+      analytics.identify('id', { name: 'Name' });
+      expect(stub.calledWith(['setCustomFields', { id: 'id', name: 'Name' }])).to.be(true);
     });
-
   });
-
 });